Update Hyperparameter_optimization_statistics.R
For applying iQC normalisation to the QC samples, it currently uses the average across all QCs (lines 506). 

If wanting to keep QC samples, but divide each in the pair (based on order) by the average across the pair, a few lines of extra code have been added here 507-513 (currently # out). It's not the most succinct code, but it works correctly.